Addressed Review Comments, New event integration dependency injection
